games_backend/app/requirements.txt
asgiref==3.4.1
click==8.0.1
fastapi==0.68.0
h11==0.12.0
httptools==0.2.0
importlib-metadata==4.6.4
psycopg2-binary==2.9.1
pychorus==0.1
pydantic==1.8.2
pydub==0.25.1
python-dotenv==0.19.0
PyYAML==5.4.1
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==9.1
zipp==3.5.0
